refactor(edge): move stacks manager to internal
