+++ http/tunnel/tunnel.go
@@ -9,6 +9,7 @@ import (
 	"github.com/portainer/agent"
 	"github.com/portainer/agent/chisel"
 	"github.com/portainer/agent/filesystem"
+	"github.com/portainer/agent/internal/edgestacks"
 )
 
 const tunnelActivityCheckInterval = 30 * time.Second
@@ -26,7 +27,7 @@ type Operator struct {
 	scheduleManager         agent.Scheduler
 	lastActivity            time.Time
 	refreshSignal           chan struct{}
-	edgeStackManager        agent.EdgeStackManager
+	edgeStacksManager       *edgestacks.Manager
 	portainerURL            string
 	endpointID              string
 	tunnelServerAddr        string
@@ -43,7 +44,7 @@ type OperatorConfig struct {
 }
 
 // NewTunnelOperator creates a new reverse tunnel operator
-func NewTunnelOperator(edgeStackManager agent.EdgeStackManager, config *OperatorConfig) (*Operator, error) {
+func NewTunnelOperator(edgeStacksManager *edgestacks.Manager, config *OperatorConfig) (*Operator, error) {
 	pollFrequency, err := time.ParseDuration(config.PollFrequency)
 	if err != nil {
 		return nil, err
@@ -63,7 +64,7 @@ func NewTunnelOperator(edgeStackManager agent.EdgeStackManager, config *Operator
 		tunnelClient:          chisel.NewClient(),
 		scheduleManager:       filesystem.NewCronManager(),
 		refreshSignal:         nil,
-		edgeStackManager:      edgeStackManager,
+		edgeStacksManager:     edgeStacksManager,
 	}, nil
 }
